refactor: re-package the project
