--- src/main/java/com/programmers/vouchermanagement/customer/CustomerRepository.java
@@ -1,7 +1,0 @@
-package com.programmers.vouchermanagement.customer;
-
-import java.util.List;
-
-public interface CustomerRepository {
-    List<Customer> findBlackCustomers();
-}
