+++ src/main/java/com/programmers/vouchermanagement/customer/controller/CustomerController.java
@@ -1,8 +1,11 @@
-package com.programmers.vouchermanagement.customer;
+package com.programmers.vouchermanagement.customer.controller;
 
 import java.util.List;
 
 import org.springframework.stereotype.Controller;
+
+import com.programmers.vouchermanagement.customer.domain.Customer;
+import com.programmers.vouchermanagement.customer.service.CustomerService;
 
 @Controller
 public class CustomerController {
