+++ src/main/java/com/programmers/vouchermanagement/customer/repository/CustomerRepository.java
@@ -1,0 +1,9 @@
+package com.programmers.vouchermanagement.customer.repository;
+
+import java.util.List;
+
+import com.programmers.vouchermanagement.customer.domain.Customer;
+
+public interface CustomerRepository {
+    List<Customer> findBlackCustomers();
+}
