--- src/main/java/com/programmers/vouchermanagement/voucher/CreateVoucherRequestDTO.java
@@ -1,6 +1,0 @@
-package com.programmers.vouchermanagement.voucher;
-
-import com.programmers.vouchermanagement.consolecomponent.CreateVoucherMenu;
-
-public record CreateVoucherRequestDTO(CreateVoucherMenu createVoucherMenu, long discountAmount) {
-}
