--- src/main/java/com/programmers/vouchermanagement/voucher/VoucherRepository.java
@@ -1,8 +1,0 @@
-package com.programmers.vouchermanagement.voucher;
-
-import java.util.List;
-
-public interface VoucherRepository {
-    void save(Voucher voucher);
-    List<Voucher> findAll();
-}
