+++ src/main/java/com/programmers/vouchermanagement/voucher/dto/CreateVoucherRequestDTO.java
@@ -1,0 +1,6 @@
+package com.programmers.vouchermanagement.voucher.dto;
+
+import com.programmers.vouchermanagement.consoleapp.menu.CreateVoucherMenu;
+
+public record CreateVoucherRequestDTO(CreateVoucherMenu createVoucherMenu, long discountAmount) {
+}
