+++ src/main/java/com/programmers/vouchermanagement/voucher/repository/VoucherRepository.java
@@ -1,0 +1,10 @@
+package com.programmers.vouchermanagement.voucher.repository;
+
+import java.util.List;
+
+import com.programmers.vouchermanagement.voucher.domain.Voucher;
+
+public interface VoucherRepository {
+    void save(Voucher voucher);
+    List<Voucher> findAll();
+}
